fix(messenger): refracto Message component
