+++ js/packages/components/chat/MultiMember.tsx
@@ -31,7 +31,8 @@ import { messenger as messengerpb } from '@berty-tech/api/index.js'
 import * as api from '@berty-tech/api/index.pb'
 
 import { ChatFooter, ChatDate } from './shared-components/Chat'
-import { Message, MessageSystemWrapper } from './shared-components/Message'
+import { Message } from './message'
+import { MessageSystemWrapper } from './message/MessageSystemWrapper'
 import BlurView from '../shared-components/BlurView'
 import { SwipeNavRecognizer } from '../shared-components/SwipeNavRecognizer'
 import AvatarGroup19 from '../main/Avatar_Group_Copy_19.png'
